Move #suppress_warnings to a namespaced Module
Adding methods to Kernel is likely to break apps, so we should scope it
to this gem only.
