--- lib/fillable-pdf/kernel.rb
@@ -1,9 +1,0 @@
-module Kernel
-  def suppress_warnings
-    original_verbosity = $VERBOSE
-    $VERBOSE = nil
-    result = yield
-    $VERBOSE = original_verbosity
-    result
-  end
-end
