+++ lib/fillable-pdf/suppress_warnings.rb
@@ -1,0 +1,11 @@
+class FillablePDF
+  module SuppressWarnings
+    def suppress_warnings
+      original_verbosity = $VERBOSE
+      $VERBOSE = nil
+      result = yield
+      $VERBOSE = original_verbosity
+      result
+    end
+  end
+end
